refactor: ルートからページ責務を分離
