+++ src/components/features/Line/useRedeemLineToken.ts
@@ -1,0 +1,6 @@
+import { useAction } from "convex/react";
+import { api } from "@/convex/_generated/api";
+
+export function useRedeemLineToken() {
+  return useAction(api.line.actions.redeemLineToken);
+}
