+++ src/components/features/StaffLegalConsent/useAcceptStaffLegalConsent.ts
@@ -1,0 +1,7 @@
+import { useMutation } from "convex/react";
+import { api } from "@/convex/_generated/api";
+
+export function useAcceptStaffLegalConsent(token: string) {
+  const accept = useMutation(api.legal.mutations.acceptStaffLegalConsent);
+  return () => accept({ token, acceptedLegal: true });
+}
