+++ src/pages/shift-board/index.tsx
@@ -1,0 +1,32 @@
+import { Flex, Spinner } from "@chakra-ui/react";
+import { useQuery } from "convex/react";
+import { api } from "@/convex/_generated/api";
+import type { Id } from "@/convex/_generated/dataModel";
+import { ShiftBoardPage } from "@/src/components/features/ShiftBoard/ShiftBoardPage";
+import { Animation } from "@/src/components/templates/Animation";
+
+type Props = {
+  recruitmentId: string;
+};
+
+export function ShiftBoardRoutePage({ recruitmentId }: Props) {
+  const data = useQuery(api.shiftBoard.queries.getShiftBoardData, {
+    recruitmentId: recruitmentId as Id<"recruitments">,
+  });
+
+  if (data === undefined) {
+    return (
+      <Flex justify="center" align="center" minH="200px">
+        <Spinner />
+      </Flex>
+    );
+  }
+
+  if (data === null) return null;
+
+  return (
+    <Animation>
+      <ShiftBoardPage data={data} recruitmentId={recruitmentId as Id<"recruitments">} />
+    </Animation>
+  );
+}
