+++ src/routes/_auth/shiftboard.$recruitmentId.tsx
@@ -1,11 +1,6 @@
-import { Flex, Spinner } from "@chakra-ui/react";
 import { createFileRoute } from "@tanstack/react-router";
-import { useQuery } from "convex/react";
-import { api } from "@/convex/_generated/api";
-import type { Id } from "@/convex/_generated/dataModel";
-import { ShiftBoardPage } from "@/src/components/features/ShiftBoard/ShiftBoardPage";
-import { Animation } from "@/src/components/templates/Animation";
 import { buildMeta } from "@/src/helpers/seo";
+import { ShiftBoardRoutePage } from "@/src/pages/shift-board";
 
 export const Route = createFileRoute("/_auth/shiftboard/$recruitmentId")({
   head: () => ({
@@ -16,23 +11,5 @@ export const Route = createFileRoute("/_auth/shiftboard/$recruitmentId")({
 
 function ShiftBoardRoute() {
   const { recruitmentId } = Route.useParams();
-  const data = useQuery(api.shiftBoard.queries.getShiftBoardData, {
-    recruitmentId: recruitmentId as Id<"recruitments">,
-  });
-
-  if (data === undefined) {
-    return (
-      <Flex justify="center" align="center" minH="200px">
-        <Spinner />
-      </Flex>
-    );
-  }
-
-  if (data === null) return null;
-
-  return (
-    <Animation>
-      <ShiftBoardPage data={data} recruitmentId={recruitmentId as Id<"recruitments">} />
-    </Animation>
-  );
+  return <ShiftBoardRoutePage recruitmentId={recruitmentId} />;
 }
