+++ src/routes/_unregistered/shifts.submit.tsx
@@ -1,17 +1,6 @@
-import { createFileRoute, useNavigate } from "@tanstack/react-router";
-import { useMutation, useQuery } from "convex/react";
-import { api } from "@/convex/_generated/api";
-import type { Id } from "@/convex/_generated/dataModel";
-import type { DayEntry } from "@/src/components/features/StaffSubmit/DayCard";
-import { ExpiredSubmitView } from "@/src/components/features/StaffSubmit/ExpiredSubmitView";
-import { ShiftSubmitPage } from "@/src/components/features/StaffSubmit/ShiftSubmitPage";
-import { NetworkErrorView } from "@/src/components/features/StaffView/NetworkErrorView";
-import { RateLimitedView } from "@/src/components/features/StaffView/RateLimitedView";
-import { StaffLayout } from "@/src/components/templates/StaffLayout";
-import { ErrorBoundary } from "@/src/components/ui/ErrorBoundary";
-import { FullPageSpinner } from "@/src/components/ui/FullPageSpinner";
+import { createFileRoute } from "@tanstack/react-router";
 import { buildMeta } from "@/src/helpers/seo";
-import { useStaffSession } from "@/src/hooks/useStaffSession";
+import { StaffShiftSubmitPage } from "@/src/pages/staff-shift-submit";
 
 export const Route = createFileRoute("/_unregistered/shifts/submit")({
   validateSearch: (search: Record<string, unknown>) => ({
@@ -25,56 +14,5 @@ export const Route = createFileRoute("/_unregistered/shifts/submit")({
 
 function ShiftSubmitRoute() {
   const { token } = Route.useSearch();
-  const state = useStaffSession(token);
-
-  if (state.status === "loading") return <FullPageSpinner />;
-  if (state.status === "rateLimited") return <RateLimitedView title="シフト提出" />;
-  if (state.status === "networkError") {
-    return (
-      <StaffLayout shopName="シフト提出">
-        <NetworkErrorView onRetry={state.retry} />
-      </StaffLayout>
-    );
-  }
-  if (state.status === "expired") return <ExpiredSubmitView shopName="シフト提出" />;
-
-  return (
-    <ErrorBoundary
-      fallback={<ExpiredSubmitView shopName="シフト提出" />}
-      onError={(error) => {
-        if (error.message?.includes("ArgumentValidationError")) {
-          state.clearSession();
-        }
-      }}
-    >
-      <ShiftSubmitContent session={state.session} />
-    </ErrorBoundary>
-  );
-}
-
-function ShiftSubmitContent({ session }: { session: { sessionToken: string; recruitmentId: string } }) {
-  const navigate = useNavigate();
-  const data = useQuery(api.shiftSubmission.queries.getSubmissionPageData, {
-    sessionToken: session.sessionToken,
-    recruitmentId: session.recruitmentId as Id<"recruitments">,
-  });
-  const submitMutation = useMutation(api.shiftSubmission.mutations.submitShiftRequests);
-
-  if (data === undefined) return <FullPageSpinner />;
-  if (data === null) return <ExpiredSubmitView shopName="シフト提出" />;
-
-  const handleSubmit = async (entries: DayEntry[], acceptedLegal?: boolean) => {
-    const requests = entries
-      .filter((e) => e.isWorking)
-      .map((e) => ({ date: e.date, startTime: e.startTime, endTime: e.endTime }));
-    await submitMutation({
-      sessionToken: session.sessionToken,
-      recruitmentId: session.recruitmentId as Id<"recruitments">,
-      requests,
-      acceptedLegal,
-    });
-    await navigate({ to: "/shifts/submit/completed" });
-  };
-
-  return <ShiftSubmitPage data={data} onSubmit={handleSubmit} />;
+  return <StaffShiftSubmitPage token={token} />;
 }
